Initial Upload
This Python file can be used to bridge Discord and Hotline. You'll need to have a Discord Application setup already to use this.

Works pretty well but looking forward to seeing improvements anyone makes. This uses the WebHook URL as well to allow you to use the Hotline username as the discord username sending the chat. Allowing for a cleaner look.